changed fabsf to fabs
updated report score

updated gamekithelper protocols and delegate methods

removed @synthesis (no longer required since Xcode 4.4) <— not true, generates warnings

iRate version update

iOS8 support

+++ MacAttack/Enemy.m
@@ -146,7 +146,7 @@
     
     nowTime = [[NSDate date] timeIntervalSince1970];
     
-    deadTime = fabsf(timeOfDeath-nowTime);
+    deadTime = fabs(timeOfDeath-nowTime);
     
 
 }
@@ -160,7 +160,6 @@
 - (void) reincarnation{}
 - (BOOL) atCenter{}
 */
-
 
 
 @end
